Optimized logging and changed motors code.
The old code to control motors seems to be modified to overcome the mistakes in the board design. Since the new board is correct, I don't see a point in keeping this extra layer of confusion. Made the code to work with the correct board design and correct config (KMS 2024). Other configs will not work anymore because of this.